+++ src/main/kotlin/com/nastyaanastasya/jetpackcompose/ui/plugin/core/BaseViewModel.kt
@@ -1,0 +1,7 @@
+package com.nastyaanastasya.jetpackcompose.ui.plugin.core
+
+import com.nastyaanastasya.jetpackcompose.ui.plugin.dependencies.Dependencies
+
+abstract class BaseViewModel {
+    protected val scope = Dependencies.ioScope
+}