+++ src/main/kotlin/com/nastyaanastasya/jetpackcompose/ui/plugin/core/PropertyKeys.kt
@@ -1,0 +1,29 @@
+package com.nastyaanastasya.jetpackcompose.ui.plugin.core
+
+object PropertyKeys {
+    const val PackageName = "PACKAGE_NAME"
+    const val ListName = "LIST_NAME"
+
+    const val UseRepPassword = "USE_REP_PASSWORD"
+    const val UseForgetPasswordText = "USE_FORGET_PASSWORD_TEXT"
+
+    const val UseOutlinedTextField = "USE_OUTLINED_TEXT_FIELD"
+
+    const val UseTrailingIcon = "USE_TRAILING_ICON"
+
+    const val UseAlreadyRegisteredText = "USE_ALREADY_REGISTERED_TEXT"
+
+    const val UseListTopAppBar = "USE_LIST_TOP_APP_BAR"
+    const val UseOneLineListItem = "USE_ONE_LINE_LIST_ITEM"
+    const val UseTwoLineListItem = "USE_TWO_LINES_LIST_ITEM"
+    const val UseThreeLineListItem = "USE_THREE_LINES_LIST_ITEM"
+    const val UseListItemWithImage = "USE_LIST_ITEM_WITH_IMAGE"
+    const val UseScrollToTopButton = "USE_SCROLL_TO_TOP_BUTTON"
+
+    const val UseTopAppBarSearchIcon = "USE_TOP_APP_BAR_SEARCH_ICON"
+
+    const val UseTopAppBarOptionsMenuIcon = "USE_TOP_APP_BAR_OPTIONS_MENU_ICON"
+
+    const val UseMenuNavigationIcon = "USE_MENU_NAVIGATION_ICON"
+    const val UseArrowBackNavigationIcon = "USE_ARROW_BACK_NAVIGATION_ICON"
+}