+++ src/main/kotlin/com/nastyaanastasya/jetpackcompose/ui/plugin/features/list/domain/ListFeatureProperties.kt
@@ -1,0 +1,10 @@
+package com.nastyaanastasya.jetpackcompose.ui.plugin.features.list.domain
+
+data class ListFeatureProperties(
+        val useScrollToTopButton: Boolean,
+        val useOneLineListItem: Boolean,
+        val useTwoLinesListItem: Boolean,
+        val useThreeLinesListItem: Boolean,
+        val useListItemWithImage: Boolean,
+        val useListTopAppBar: Boolean
+)